Deleting Files

+++ controllers/admin.js
@@ -1,5 +1,6 @@
 const { validationResult } = require("express-validator");
 
+const fileHelper = require("../util/file");
 const Product = require("../models/product");
 
 exports.getProducts = (req, res, next) => {
@@ -155,7 +156,8 @@ exports.postEditProduct = (req, res, next) => {
       product.price = updatedPrice;
       product.description = updatedDescription;
       if (image) {
-        product.imageUrl = image.path;        
+        fileHelper.deleteFile(product.imageUrl);
+        product.imageUrl = image.path;
       }
       return product.save().then((result) => {
         res.redirect("/admin/products");
@@ -170,7 +172,15 @@ exports.postEditProduct = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
+
+  Product.findById(prodId)
+    .then((product) => {
+      if (product == null) {
+        return next(new Error("Product not found."));
+      }
+      fileHelper.deleteFile(product.imageUrl);
+      return Product.deleteOne({ _id: prodId, userId: req.user._id });
+    })
     .then(() => {
       console.log("DESTROYED PRODUCT");
       res.redirect("/admin/products");
